requirements.txt
certifi==2022.5.18.1
charset-normalizer==2.0.12
click==8.1.3
dominate==2.6.0
Flask==2.1.2
Flask-Bootstrap4==4.0.2
Flask-WTF==1.0.1
idna==3.3
importlib-metadata==4.11.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
Pillow==9.1.1
pypng==0.0.21
PyQRCode==1.2.1
reportlab==3.6.9
requests==2.27.1
urllib3==1.26.9
visitor==0.1.3
Werkzeug==2.1.2
WTForms==3.0.1
zipp==3.8.0
